updated scripts

+++ docker/scripts/run.dev.sh
@@ -3,6 +3,12 @@
 SCRIPTDIR=$(dirname "${BASH_SOURCE[0]}")
 WORKDIR=$SCRIPTDIR/../..
 
+if [ "$1" != "cpu" ] && [ "$1" != "gpu" ]; then
+	echo "missing compute option CPU or GPU"
+	echo "example usage: ./scripts/build.sh cpu"
+	exit 1
+fi
+
 docker run --rm -it \
 	-v $PWD/$WORKDIR/Output:/workdir/model \
 	-v $PWD/$WORKDIR/docker/context/flaskapi:/workdir/flaskapi \
@@ -11,7 +17,7 @@ docker run --rm -it \
 	-e PORT=80 \
 	-p 3000:80 \
 	--name vfastrcnn \
-	user1m/vott-reviewer-cntk:cpu \
+	user1m/vott-reviewer-cntk:$1 \
 	bash -c "/scripts/init.sh"
 
 # -v $PWD/$WORKDIR/api:/workdir/api \
